Move pin entry into its own package
