+++ padlock-lock/build.gradle
@@ -73,7 +73,7 @@ dependencies {
   // Mockito for unit tests
   testCompile "org.mockito:mockito-core:$mockito"
 
-  debugCompile project(path: ":padlock-pin", configuration: "debug")
-  releaseCompile project(path: ":padlock-pin", configuration: "release")
+  debugCompile project(path: ":padlock-base", configuration: "debug")
+  releaseCompile project(path: ":padlock-base", configuration: "release")
 }
