+++ padlock/src/main/java/com/pyamsoft/padlock/list/LockListComponent.java
@@ -17,7 +17,7 @@
 package com.pyamsoft.padlock.list;
 
 import com.pyamsoft.padlock.iconloader.AppIconLoaderModule;
-import com.pyamsoft.padlock.pin.MasterPinModule;
+import com.pyamsoft.padlock.lock.master.MasterPinModule;
 import com.pyamsoft.padlock.service.LockServiceStateModule;
 import dagger.Subcomponent;
