+++ padlock/src/main/java/com/pyamsoft/padlock/list/LockListFragment.java
@@ -41,11 +41,11 @@ import com.pyamsoft.padlock.Injector;
 import com.pyamsoft.padlock.PadLock;
 import com.pyamsoft.padlock.R;
 import com.pyamsoft.padlock.databinding.FragmentLockListBinding;
-import com.pyamsoft.padlock.lock.PinEntryDialog;
+import com.pyamsoft.padlock.pin.PinEntryDialog;
 import com.pyamsoft.padlock.main.MainActivity;
 import com.pyamsoft.padlock.model.AppEntry;
 import com.pyamsoft.padlock.onboard.list.OnboardListDialog;
-import com.pyamsoft.padlock.pin.MasterPinSubmitCallback;
+import com.pyamsoft.padlock.lock.master.MasterPinSubmitCallback;
 import com.pyamsoft.padlock.service.PadLockService;
 import com.pyamsoft.pydroid.ActionSingle;
 import com.pyamsoft.pydroid.design.fab.HideScrollFABBehavior;
